music bot/requirements.txt
discord.py[voice]
yt_dlp
ffmpeg-python
python-dotenv
pynacl
